[Refactor] AppDeveloperView, AppRatingView, AppSelfRatingView 하위 뷰로 분리
- 앱 개발자 영역, 앱 리뷰 및 평가 영역, 사용자 앱 리뷰 및 평가 영역 하위 뷰로 분리
